% Generated by build script
\name{regs_comment}
\alias{regs_comment}
\title{Fetch a single comment by ID}
\usage{
regs_comment()
}
\arguments{
\item{comment_id}{Comment ID}
\item{api_key}{API key}
}
\value{tibble: one row with full comment details including text}
\description{Fetch a single comment by ID}
