% Generated by build script
\name{regs_comments}
\alias{regs_comments}
\title{Search public comments on regulations}
\usage{
regs_comments()
}
\arguments{
\item{query}{Search term}
\item{agency}{Agency acronym filter}
\item{document_id}{Filter to comments on a specific document}
\item{posted_date_from}{Start date}
\item{posted_date_to}{End date}
\item{api_key}{API key}
\item{max_results}{Max results (default 100)}
}
\value{tibble: id, title, agencyId, postedDate, comment text, ...}
\description{Search public comments on regulations}
