% Generated by build script
\name{regs_dockets}
\alias{regs_dockets}
\title{Search regulatory dockets}
\usage{
regs_dockets()
}
\arguments{
\item{query}{Search term}
\item{agency}{Agency acronym filter}
\item{docket_type}{"Rulemaking" or "Nonrulemaking"}
\item{api_key}{API key}
\item{max_results}{Max results (default 50)}
}
\value{tibble: id, title, agencyId, docketType, ...}
\description{Search regulatory dockets}
